Crawl Local Library
Successfully crawls a local library and returns a HashSet of all
extracted MediaFile structs.

For future note, the MediaFile.id() function is the main bottle
neck for get_local_media_library(). It is the difference between
~100ms and 2000ms, so it makes things at least an order of
magnitude slower. What _specifically_ about it that makes it slow
remains to be seen.